Remove unnecessary imports and functions
