+++ gnn-asset-clustering/.gitattributes
@@ -1,0 +1,1 @@
+dataset/ports.txt filter=lfs diff=lfs merge=lfs -text
