--- gnn-asset-clustering/training-tuning-inference/silhouette_analysis.py
@@ -1,80 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from sklearn.metrics import silhouette_score, silhouette_samples
-import matplotlib.ticker as ticker
-import numpy as np
-
-
-def silhouette_plot(X, model, ax, colors):
-    y_lower = 10
-    y_tick_pos_ = []
-    sh_samples = silhouette_samples(X, model.labels_)
-    sh_score = silhouette_score(X, model.labels_)
-
-    for idx in range(model.n_clusters):
-        values = sh_samples[model.labels_ == idx]
-        values.sort()
-        size = values.shape[0]
-        y_upper = y_lower + size
-        ax.fill_betweenx(np.arange(y_lower, y_upper), 0, values, facecolor=colors[idx], edgecolor=colors[idx])
-        y_tick_pos_.append(y_lower + 0.5 * size)
-        y_lower = y_upper + 10
-
-    ax.axvline(x=sh_score, color="red", linestyle="--", label="Avg Silhouette Score")
-    ax.set_title("Silhouette Plot for {} clusters".format(model.n_clusters))
-    l_xlim = max(-1, min(-0.1, round(min(sh_samples) - 0.1, 1)))
-    u_xlim = min(1, round(max(sh_samples) + 0.1, 1))
-    ax.set_xlim([l_xlim, u_xlim])
-    ax.set_ylim([0, X.shape[0] + (model.n_clusters + 1) * 10])
-    ax.set_xlabel("silhouette coefficient values")
-    ax.set_ylabel("cluster label")
-    ax.set_yticks(y_tick_pos_)
-    ax.set_yticklabels(str(idx) for idx in range(model.n_clusters))
-    ax.xaxis.set_major_locator(ticker.MultipleLocator(0.1))
-    ax.legend(loc="best")
-    return ax
-
-
-'''
-# Plot for multiple ks
-k_max = 6
-ncols = 3
-# calculate number of rows
-nrows = k_max // ncols + (k_max % ncols > 0)
-# set figure size
-fig = plt.figure(figsize=(15,15), dpi=200)
-
-for k in range(2,k_max+1):
-    
-    # Apply clustering for k clusters
-    km = KMeans(n_clusters=k, 
-                max_iter=300, 
-                tol=1e-04, 
-                init='k-means++', 
-                n_init=10, 
-                random_state=42, 
-                algorithm='auto')
-
-    km_fit = km.fit(X_std)
-    
-    # create chart
-    ax = plt.subplot(nrows, ncols, k-1)
-    silhouette_plot(X_std, km_fit,ax, cluster_colors)
-
-fig.suptitle("Silhouette plots", fontsize=18, y=1)
-plt.tight_layout()
-plt.show()
-'''
